feat: 整合 UI 設定提供者以支持動態字體大小；更新多個元件以使用新的字體大小邏輯；更新版本號至 0.1.7+28

+++ lib/widgets/attendance_punch_card.dart
@@ -44,7 +44,7 @@ class AttendancePunchCard extends StatelessWidget {
                   child: Text(
                     displayName,
                     style: TextStyle(
-                      fontSize: 18 * uiSettings.fontSizeScale,
+                      fontSize: (18 * uiSettings.fontSizeScale).toDouble(),
                       fontWeight: FontWeight.bold,
                     ),
                   ),
